requirements.txt
livekit-agents==0.12.15
python-dotenv
openai
